+++ config/forex.php
@@ -3,4 +3,5 @@
 return [
     'email_recipients' => env('FOREX_EMAIL_RECIPIENTS', ''),
     'seed_default_symbols' => env('FOREX_SEED_DEFAULT_SYMBOLS', false),
+    'seed_admin_user' => env('FOREX_SEED_ADMIN_USER', false),
 ];